+++ pages/api/verifier/presentation-response.ts
@@ -9,7 +9,7 @@ type Data = {
 
 export default function handler(
   req: NextApiRequest,
-  res: NextApiResponse<Data>,
+  res: NextApiResponse<Data>
 ) {
   res.status(200).json({
     status: 0,
